+++ Reto/SituacionProblema_A01721732/src/Episodio.cpp
@@ -7,7 +7,7 @@ Episodio::Episodio()
     temporada = 0;
     calificacion = 0;
 }
-Episodio::Episodio(string _titulo, int _temporada, int _calificacion)
+Episodio::Episodio(string _titulo, int _temporada, double _calificacion)
 {
     titulo = _titulo;
     temporada = _temporada;
@@ -23,7 +23,7 @@ void Episodio::setTemporada(int _temporada)
 {
     temporada = _temporada;
 }
-void Episodio::setCalificacion(int _calificacion)
+void Episodio::setCalificacion(double _calificacion)
 {
     calificacion = _calificacion;
 }
@@ -37,7 +37,7 @@ int Episodio::getTemporada()
 {
     return temporada;
 }
-int Episodio::getCalificacion()
+double Episodio::getCalificacion()
 {
     return calificacion;
 }
